fix: Include <cstdlib> for std::rand() function
- Added the <cstdlib> library to ensure std::rand() is available.
- Ensured compatibility and proper usage of the standard library function.

+++ Module_07/ex01/src/main.cpp
@@ -2,7 +2,7 @@
 #include "../include/test.hpp"
 
 static const TestFunction testFunctions[] = {
-    {"int", testInt}, {"float", testDouble}, {"string", testString}, {"all", allTest}};
+    {"int", testInt}, {"double", testDouble}, {"string", testString}, {"all", allTest}};
 
 void allTest(bool skip) {
     (void)skip;
